% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat_files.R
\name{Concat_files}
\alias{Concat_files}
\title{Combines data from files produced by \code{\link{LDnClusteringEL}}}
\usage{
Concat_files(LDnCl_out)
}
\arguments{
\item{LDnCl_out}{Path to folder containing output from \code{\link{LDnClusteringEL}}}
}
\value{
Returns concatenated file with the same data as in the individual \code{.rds} files. Can be used with e.g. \code{\link{emmax_group}} as with output produced by \code{\link{LDnClusteringEL}}
}
\description{
\code{\link{LDnClusteringEL}} produces one \code{.rds} file for each chromosome. This function concatenates the data from each \code{.rds} file into a single file
}
\examples{
\dontrun{
LDnCL_combined <- Concat_files("path/to/rds/files")
}                   
}
\references{
Kemppainen, P., Knight, C. G., Sarma, D. K., Hlaing, T., Prakash, A., Maung Maung, Y. N., Walton, C. (2015). Linkage disequilibrium network analysis (LDna) gives a global view of chromosomal inversions, local adaptation and geographic structure. Molecular Ecology Resources, 15(5), 1031-1045. https://doi.org/10.1111/1755-0998.12369\cr
\cr
Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li \email{lizitong1985@gmail.com}
}
\keyword{Linkage}
\keyword{clustering,}
\keyword{complexity}
\keyword{disequilibrium}
\keyword{network}
\keyword{reduction}
