% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBranches.R
\name{extractBranches}
\alias{extractBranches}
\title{Extracts LD clusters for linkage disequilibrium network analysis (LDna) using only |E|min, i.e. considers each branch as LD-cluster.}
\usage{
extractBranches(ldna, min.edges = 20, merge.min = 0.8, plot.tree = TRUE)
}
\arguments{
\item{ldna}{Output from \code{\link{LDnaRaw}}}

\item{min.edges}{Minimum number of edges for a cluster that is shown as a branch in a tree.}

\item{merge.min}{Is the correlation threshold at which a clade is considered a separate LD-cluster, even if it contains more than two branches.}

\item{plot.tree}{If \code{TRUE} (default), plots tree.}

\item{LDmat}{Same LDmat as used for \code{\link{LDnaRaw}}}
}
\value{
Returns a list containing a vectors of locus names belonging to a given cluster (branch)
}
\description{

}
\details{
If \code{plot.tree} and \code{plot.graph} are set to \code{TRUE}, \code{extractClusters} plots two graphs (default). The first shows all \eqn{\lambda}-values oredered from low to high and indicates which values are above \eqn{\lambda_{lim}}. Values corresponding to \emph{"selected outlier clusters", SOCs} are indicated in red and values corresponding to \emph{COCs} are indiced in blue. A \emph{COC} is defined as any \emph{OC} that contains loci from an \emph{OC} already extracted at a higher LD threshold. The second graph gives the tree illustrating cluster merger with decreasing LD threshold where nodes represent and node distance gives the LD threholds at which these events occur. Branches corresponding to \emph{SOCs} are indicated in red and those corresponding to \emph{COCs} are indiced in blue (if \code{rm.COCs=FALSE}).

Branch traversal means that as long as a branch contains outliers, the outlier clusters (SOC or COC) that is closest to the base of the branch is selected as the outlier cluster. This ensures that a wide range threshold values for \eqn{\lambda} gives similar results.
}
\examples{
data(LDna)

ldna <- LDnaRaw2(r2.baimaii_subs)

par(mfcol=c(1,2))
clusters <- extractBranches(ldna,min.edges=20,merge.min=0.8, plot.tree=TRUE) # default values
clusters
}
\seealso{
\code{\link{LDnaRaw}}, \code{\link{summaryLDna}} and \code{\link{plotLDnetwork}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, Christopher Knight \email{Chris.Knight@manchester.ac.uk}
}
\keyword{extractBranches}
