\docType{package}
\name{LDna}
\alias{LDna}
\alias{LDna-package}
\title{Linkage disequilibrium network analyses (LDna)}
\description{
Linkage disequilibrium anlyses is an exploratory tool for
population genomic data that can identify distinct
population genetic signals and identify loci causing them.
}
\details{
The main pipeline is to produce raw data using:
\code{\link{LDnaRaw}}, extract clusters using:
\code{\link{extractClusters}}, summerise data using:
\code{\link{summaryLDna}} and exploring networks using
\code{\link{plotLDnetwork}}.
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and
Christopher Knight Christopher Knight,
\email{Chris.Knight@manchester.ac.uk}
}

