% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDnaBinning.R
\name{LDnBin}
\alias{LDnBin}
\title{Linkage disequilibrium (LD) network binning}
\usage{
LDnBin(snp, mapsnp, nSNPs = 1000, w1 = 10, w2 = 100,
  LD_threshold1 = 0.1, LD_threshold2 = 0.3, PC_threshold = 0.8,
  verbose = FALSE, plot.tree = FALSE, mc.cores = 1, plot.network = NULL,
  threshold_net = 0.9)
}
\arguments{
\item{snp}{A matrix with individuals as rows and bi-allelic SNP genotypes as columns. SNPs must be coded as 0,1 or 2 for the number of alleles of the reference nucleotide.}

\item{mapsnp}{A matrix with each row correspoinding a column in \code{snp}, column one corresponding to chromosome or linkage group and column two corresponding to physical position in the genome.}

\item{nSNPs}{Desired number of SNPs per bin.}

\item{w1}{Window size for defining putative recombination hotspots.}

\item{w2}{Window size for estimating LD values.}

\item{LD_threshold1}{Minimum LD value within a cluster}

\item{LD_threshold2}{Minimum median LD within each cluster, and for the second step of clustering}

\item{PC_threshold}{Minimum cummulative amount of genetic variation explained by extracted PCs}

\item{verbose}{More detailed information of progress is outputted}

\item{plot.tree}{Plot tree. Default \code{FALSE}}

\item{mc.cores}{number of cores for mclapply}

\item{plot.network}{File name for plotting network. If \code{NULL} (default) no network is plotted.}

\item{threshold_net}{Threshold for edges when plotting network.}
}
\value{
Returns a list of three objects. \code{mapbin} contains map information for each PC extracted from the data, PCbin=PCbin, bin_snps=bin_snps \cr
test
}
\description{
Finds clusters of loci connected by high LD within a sliding window summerizes genetic variation using principal component analyses
}
\details{
If \code{plot.tree} and \code{plot.graph} are set to \code{TRUE}, \code{extractClusters} plots two graphs (default). The first shows all \eqn{\lambda}-values oredered from low to high and indicates which values are above \eqn{\lambda_{lim}}. Values corresponding to \emph{"selected outlier clusters", SOCs} are indicated in red and values corresponding to \emph{COCs} are indiced in blue. A \emph{COC} is defined as any \emph{OC} that contains loci from an \emph{OC} already extracted at a higher LD threshold. The second graph gives the tree illustrating cluster merger with decreasing LD threshold where nodes represent and node distance gives the LD threholds at which these events occur. Branches corresponding to \emph{SOCs} are indicated in red and those corresponding to \emph{COCs} are indiced in blue (if \code{rm.COCs=FALSE}).
}
\examples{
# To come
}
\seealso{
\code{\link{LDnaRaw}}, \code{\link{summaryLDna}} and \code{\link{extractClusters}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li
}
\keyword{LDnB}
