% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDnClustering.R
\name{LDnClustering}
\alias{LDnClustering}
\title{Linkage disequilibrium (LD) network clustering}
\usage{
LDnClustering(snp, map, nSNPs = 1000, w1 = 10, w2 = 100,
  LD_threshold1 = 0.1, LD_threshold2 = 0.3, PC_threshold = 0.8,
  verbose = TRUE, plot.tree = FALSE, mc.cores = 1, plot.network = NULL,
  threshold_net = 0.9)
}
\arguments{
\item{snp}{A matrix with individuals as rows and bi-allelic SNP genotypes as columns. SNPs must be coded as 0,1 or 2 for the number of alleles of the reference nucleotide.}

\item{map}{A matrix with each row correspoinding a column in \code{snp}, column one corresponding to chromosome or linkage group and column two corresponding to physical position in the genome.}

\item{nSNPs}{Desired number of SNPs per window.}

\item{w1}{Window size for defining putative recombination hotspots. Default is 10.}

\item{w2}{Window size for estimating LD values (smaller is faster but may break up large clusters caused e.g. by inversion polymorphism).}

\item{LD_threshold1}{Minimum LD value within a cluster}

\item{LD_threshold2}{Minimum median LD within each cluster}

\item{PC_threshold}{Minimum cummulative amount of genetic variation explained by extracted PCs}

\item{verbose}{More detailed information of progress is printed on screen}

\item{plot.tree}{Plot complete linkage tree.}

\item{mc.cores}{Number of cores for mclapply}

\item{plot.network}{File name for plotting network. If \code{NULL} (default) no network is plotted.}

\item{threshold_net}{Threshold for edges when plotting network.}
}
\description{
Finds clusters of loci connected by high LD within non-overlapping windows and summerises this information using Principal Component Analysis, to be used in Genome Wide Association (GWA) analyses.
}
\details{
Uses complete linkage clustering within non-overlapping windows of SNPs to find clusters of SNPs connected by high LD. Window breakpoints are placed where r^2 between adjacent SNPs within a window (defined by \code{w1}) is below \code{LD_threshold1}.
}
\examples{
## Arabidopsis sample data (1000 first SNPs, Chr 1)
data(LDna)

data_0.1_0.3 <- LDnClustering(snp=snp_ara, 
                                  map=map_ara, 
                                  nSNPs = 1000, 
                                  w2 = 100, 
                                  LD_threshold1 = 0.1, 
                                  LD_threshold2 = 0.3, 
                                  verbose = TRUE)
data_0.3_0.5 <- LDnClustering(snp=snp_ara, 
                                  map=map_ara, 
                                  nSNPs = 1000, 
                                  w2 = 100, 
                                  LD_threshold1 = 0.3, 
                                  LD_threshold2 = 0.5, 
                                  verbose = TRUE)
}
\references{
Kemppainen, P., Knight, C. G., Sarma, D. K., Hlaing, T., Prakash, A., Maung Maung, Y. N., Walton, C. (2015). Linkage disequilibrium network analysis (LDna) gives a global view of chromosomal inversions, local adaptation and geographic structure. Molecular Ecology Resources, 15(5), 1031-1045. https://doi.org/10.1111/1755-0998.12369\cr
\cr
Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
\seealso{
\code{\link{emmax_group}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li \email{lizitong1985@gmail.com}
}
\keyword{Linkage}
\keyword{clustering,}
\keyword{complexity}
\keyword{disequilibrium}
\keyword{network}
\keyword{reduction}
