% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDnClusteringEL.R
\name{LDnClusteringEL}
\alias{LDnClusteringEL}
\title{Linkage disequilibrium (LD) network clustering, with pre-calculated LD values in edgelists as input}
\usage{
LDnClusteringEL(
  EL_path = "./LD_EL",
  out_folder = "LDnCL_out",
  LD_column = 3,
  nSNPs = 1000,
  w1 = 10,
  LD_threshold1 = 0.5,
  LD_threshold2 = 0.8,
  PC_threshold = 0,
  mc.cores = 1,
  plot.network = NULL,
  threshold_net = 0.9,
  to_do = NULL,
  min.cl.size = 2
)
}
\arguments{
\item{EL_path}{Path to folder with edge lists of pairwise LD values. Typically you can restrict the window size to ~100 consequtive SNPs to substantially reduce the data sets and computation times. This is essentially the parameter \code{W2}. Default is "./LD_EL"}

\item{out_folder}{Is the folder where all output is created. Use \code{Concat_files(out_folder)} that takes the folder names as argument and concatenates all results. If this folder does not exist, it is created.}

\item{LD_column}{Is the column in the edge lists that contains the LD estimate.}

\item{nSNPs}{Desired number of SNPs per window.}

\item{w1}{Window size for defining putative recombination hotspots. Default is 10.}

\item{LD_threshold1}{Minimum LD value within a cluster}

\item{LD_threshold2}{Minimum median LD within each cluster}

\item{PC_threshold}{Minimum cummulative amount of genetic variation explained by extracted PCs}

\item{mc.cores}{Number of cores for mclapply}

\item{plot.network}{File name for plotting network. If \code{NULL} (default) no network is plotted.}

\item{threshold_net}{Threshold for edges when plotting network}

\item{to_do}{Vector with indexes for files in folder specified by \code{EL_path} that still need to be done. If \code{NULL} (default), LDn-clustering will be performed for all data files in  \code{out_folder}}

\item{min.cl.size}{If you only want to keep clusters of certain size (in terms of numbers of loci), this can be specified here. It turns out that by excluding all singleton clusters (\code{min.cl.size=2}), not much information is lost, especially for outlier analyses and data with high levels of noise (e.g. small samplesizes).}
}
\value{
Returns a list of five objects. These are saved as separate .rds files in the folder "LDnCl_out" (which is created if it does not already exist). If something goes wrong you can specify \code{to_do} with indexes for files in folder specified by \code{EL_path} that still need to be done. 
\code{cluster_summary} is a data frame that contains most of the relevant information for each cluster, inluding summary statistics (Median LD, MAD etc, see below). 
\code{cluster_PCs} includes a matrix individuals (rows) and PCs (columns) corresponding to each row in \code{cluster_summary}. This is what you use for subsequent GWA analyses. 
The third object (\code{clusters}) contains a list of locus names (chr:position; each entry corresponding to a row in \code{cluster_summary} for entries where PC=1). 
The fourth and fifth objects are files matrixes similar to \code{cluster_PCs} except there will be one column per cluster with either the SNP with the highest intra cluster median LD ('maximally connected SNP', MCL) or the consensus SNP (Cons), respectively. 
See example code for details. These last two files are expected to work better when LD thresholds are higher and SNPs in clusters or more strongly correlated, but no simualtions have been done to test this yet. 
\cr 
\cr
Each .rds file contains only information for each chromosome but this data can be concatenated into a single file using \code{\link{Concat_files}}. 
\cr
The columns in file \code{cluster_summary} are:
\item{Chr}{Chromosome or linkage group identifer}
\item{Window}{Window identifier, recycled among chromosomes}
\item{Pos}{Mean position of SNPs in a cluster}
\item{Min}{Most downstream position of SNPs in a cluster}
\item{Max}{Most upstream position of SNPs in a cluster}
\item{Range}{Max-Min}
\item{nSNPs}{Number of SNPs in the cluster}
\item{Median}{Median pair wise LD between loci in a cluster}
\item{MAD}{Median absolute deviation among pair wise LD-values between loci in a cluster}
\item{PC}{PC identifier (first PC explaining most of the genetic variation)}
\item{PVE}{Cummulative proportion of variation explained by each PC}
\item{Grp}{Unique group identifier for PCs extracted from each cluster. Used by \code{\link{emmax_group}}}
}
\description{
Finds clusters of loci connected by high LD within non-overlapping windows and summerises this information using Principal Component Analysis, to be use in Genome Wide Association (GWA) analyses. Takes pre-calculated LD edge lists as input.
}
\details{
Uses complete linkage clustering within non-overlapping windows of SNPs (within chromosomes) to find groups of correlated SNPs connected by high LD. 

\cr
Window breakpoints are placed where r^2 between adjacent SNPs within a window (defined by \code{w1}) is below \code{LD_threshold1}. 
\code{nSNPs} determines the goal size for each window; smaller windows will produce faster compuation times. \cr
\cr
First, a graph object is created using all pairwise r^2 values between SNPs within each window, and edges (representing r^2 values) < LD_threshold1 are removed. This breaks the initial graph into sub-graphs. 
Removing edges like this and decomposing the graph is equivalent to a \code{sinlge linkage} clustering algorithm, with the draw back that clusters formed via single linkage clustering may be forced together due to single elements being close to each other, even though many of the elements in each cluster may be very distant to each other ('chaining-phenomenon').
For the (potentially long and elongated) sub-clusters produced in step one, a \code{complete linkage clustering} algorithim is performed (all edged need to be connected by a threshold value of r^2, resulting in more compact and highly connected clusters). \cr
\cr
With the \code{complete linkage clustering} trees clades with median LD>\code{LD_threshold2} are are recursively found. 
Thus \code{LD_threshold1} is typically set to lower than \code{LD_threshold1} and aims to break the initial network into sub-clusters, thus increasing compuational speed (i.e. it would also be possible to produce a complete linkage clustering tree for the whole window and then recursively find the relevant clades to extract). \cr
\cr
From each cluster as many Principal Components (\code{PCs}) as is required to explain >= \code{PC_threshold} are extracted. 
To increase compuational speed for large windows, or if a physical limit to how large clusters one is intrested in, window size (\code{w2}) for estimating r^2 can also be set to something less than the window size. This will in practice limit size of the clusters to (\code{w2}).
This algorithm is performed by chromosome and window, and can thus be parallelized (\code{mc.cores}).\cr
\cr
It is also possible to plot LD-networks for different cluster solutions. For this you can specify a path for where the figure will be saved (\code{plot.network}). 
The parameter \code{threshold_net} determines the minimum value for each edge that is allowed in the network (i.e is equivalent to a single linkage clustering network). 
Each LD-cluster has a unique (arbitrary) color combination and black vertices reprsent loci that are not part of any cluster, i.e. should be considered independently in GWA analsyes. 
Minimum for \code{plot.network} is './', which exports figure to the working directory. Only works for data sets with up to 1000 SNPs and it is recommended that \code{threshold_net} is ~LD2.
}
\examples{
## example run, works basically the same as 'LDnClustering' but instead of defining 'snp' and 'map' files, you have to specify the path to a folder with the pre-calculated edge lists.

\dontrun{
library(parallel)
## first create a folder with edge lists for two chromosomes 
data("LDna")
system("mkdir LD_EL")
write.table(el[[1]], col.names = TRUE, row.names = FALSE, quote = FALSE, file="./LD_EL/chr1.el")
write.table(el[[2]], col.names = TRUE, row.names = FALSE, quote = FALSE, file="./LD_EL/chr2.el")


LDnClusteringEL(EL_path = "./LD_EL", nSNPs=1000, w1=10, out_folder="LDnCL_out_mc", LD_column = 4, LD_threshold1 = 0.5, LD_threshold2 = 0.8, PC_threshold=0, mc.cores=4, plot.network=NULL, to_do=NULL, threshold_net=0.9, min.cl.size=2)
LDn_Cl_data <- Concat_files("LDnCL_out_mc")
}
                  
}
\references{
Kemppainen, P., Knight, C. G., Sarma, D. K., Hlaing, T., Prakash, A., Maung Maung, Y. N., Walton, C. (2015). Linkage disequilibrium network analysis (LDna) gives a global view of chromosomal inversions, local adaptation and geographic structure. Molecular Ecology Resources, 15(5), 1031-1045. https://doi.org/10.1111/1755-0998.12369\cr
\cr
Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
\cr
Fang, B., Kemppainen, P., Momigliano, P., Merila, J. Oceans apart: heterogeneous patterns of parallel evolution sticklebacks. https://www.biorxiv.org/content/10.1101/826412v1
}
\seealso{
\code{\link{emmax_group}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li \email{lizitong1985@gmail.com}
}
\keyword{Linkage}
\keyword{clustering,}
\keyword{complexity}
\keyword{disequilibrium}
\keyword{network}
\keyword{reduction}
