% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ara.R
\docType{data}
\name{map_ara}
\alias{map_ara}
\title{\emph{Arabidopsis thaliana} map data (subset)}
\usage{
map_ara
}
\description{
map information for first 1000 SNPs (Chr 1) from \emph{Arabidopsis thaliana} as used in Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
