\docType{data}
\name{r2.baimaii_subs}
\alias{r2.baimaii_subs}
\title{Matrix of pairwise r-squared values}
\usage{
r2.baimaii_subs
}
\description{
Upper diagonal pairwise matrix of r-squared values between
957 SNPs from 91 wild-caught individuals of the SE Asian
mosquito malaria vector, \emph{Anopheles baimaii}.
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}

