% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_ara.R
\docType{data}
\name{snp_ara}
\alias{snp_ara}
\title{\emph{Arabidopsis thaliana} SNP data (subset)}
\usage{
snp_ara
}
\description{
first 1000 SNPs (Chr 1) from \emph{Arabidopsis thaliana} as used in Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
