% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmax_group.R
\name{emmax_group}
\alias{emmax_group}
\title{Modification of Emmax to be used for Genome Wide Association (GWA) analyses using Groups of SNPs connected by high LD.}
\usage{
emmax_group(Y, X, Grp, K, B = NULL)
}
\arguments{
\item{Y}{Phenotypic value}

\item{X}{Original SNP data used for by \code{\link{LDnClustering}} (matrix with SNPs as columns and individuals as rows)}

\item{Grp}{Index indicating which LD-cluster SNPs belong to. Produced by \code{\link{LDnClustering}}.}

\item{K}{Kinship matrix, as produced e.g by \code{A.mat}, package \code{rrBLUP}}

\item{B}{Number of permutations. If \code{NULL}, permutation is not performed}
}
\value{
See reference
}
\description{
This function is not meant to be a part of LDna but is provided here to show an example of how output from \code{\link{LDnClustering}} can be used for GWA studies. Potentially there will be a separate R-package in the future that focuses on GWA. For now we refer to code provided with the reference, see below.
}
\details{
See reference
}
\examples{
## see \code{\link{LDnClustering}}
}
\references{
Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
\seealso{
\code{\link{LDnClustering}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li \email{lizitong1985@gmail.com}
}
\keyword{GWAS,}
\keyword{Genome}
\keyword{Linkage}
\keyword{analyses,}
\keyword{association}
\keyword{clustering}
\keyword{disequilibrium,}
\keyword{network,}
\keyword{wide}
