% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBranches.R
\name{extractBranches}
\alias{extractBranches}
\title{Extracts LD clusters for linkage disequilibrium network analysis (LDna).}
\usage{
extractBranches(
  ldna,
  min.edges = 200,
  merge.min = 0.8,
  plot.tree = TRUE,
  cores = 10
)
}
\arguments{
\item{ldna}{Output from \code{\link{LDnaRaw2}}}

\item{min.edges}{Minimum number of edges for a cluster that is shown as a branch in a tree.}

\item{merge.min}{Is the correlation threshold at which a clade is considered a separate LD-cluster, even if it contains more than two branches.}

\item{plot.tree}{If \code{TRUE} (default), plots tree.}

\item{cores}{Number of cores to use.}
}
\value{
Returns a list containing a vectors of locus names belonging to a given cluster (branch)
}
\description{
Finds clusters of highly correlated loci (LD-clusters) by considering each branch in single-linkage clustering trees as a separate LD-cluster. The single-linkage can be plotted with the brances representing LD-clusters indicated.
}
\details{
In this implementation of defining LD-clusters, only the parameter \code{min.edges} needs to be specified. Low values of \code{min.edges} lead to "bushes" with many small "twigs" in the tree, each corresponding to a separate LD-cluster. These clusters contain fewer but more highly correlated sets of loci. Conversely, high values lead to fewer "thick" branches i.e. larger (in terms of the number of loci) but less strongly correlated LD-clusters.
\cr
\cr
The single-linkage clustering tree visualizing branches/LD-clusters can be printed by \code{plot.tree}=TRUE. 
\cr
\cr
With low values for \code{min.edges} an additional parameter \code{merge.min} can be used to group "branches" that merge above this threshold into a single LD-cluster; clusters that merge at high LD-threshold are so correlated, it makes no sense to regard them as separate LD-clusters.

This function works only with \code{\link{LDnaRaw2}} as shown in the examples
}
\examples{
data(LDna)

ldna <- LDnaRaw2(r2.baimaii_subs)

clusters <- extractBranches(ldna,min.edges=20,merge.min=0.8, plot.tree=TRUE) # default values
clusters

# with lower threshold for  \code{min.edges}
clusters <- extractBranches(ldna,min.edges=5,merge.min=0.8, plot.tree=TRUE) # default values
clusters

# with lower threshold for  \code{min.edges} and 
clusters <- extractBranches(ldna,min.edges=5,merge.min=0.2, plot.tree=TRUE) # default values
clusters
}
\seealso{
\code{\link{LDnaRaw2}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, Christopher Knight \email{Chris.Knight@manchester.ac.uk}
}
\keyword{extractBranches}
