% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClusters.R
\name{extractClusters}
\alias{extractClusters}
\title{Extracts LD clusters for linkage disequilibrium network analysis (LDna)}
\usage{
extractClusters(ldna, min.edges = 20, phi = 2, lambda.lim = NULL,
  rm.COCs = TRUE, extract = TRUE, plot.tree = TRUE, plot.graph = TRUE,
  LD_threshold1 = NULL, LD_threshold2 = NULL, extract.fun = function(ldna,
  LD_threshold1) {     colnames(ldna$clusterfile)[c(0, ldna$lambda_min$V1) >
  LD_threshold1 & c(0, ldna$lambda_min$V2) < LD_threshold1] })
}
\arguments{
\item{ldna}{Output from \code{\link{LDnaRaw}}}

\item{min.edges}{Minimum number of edges for a cluster that is shown as a branch in a tree.}

\item{phi}{Controls \eqn{\lambda_{lim}} which sets the threshold above which \eqn{\lambda} are considered as outliers. Default is two, values below this are not recommended.}

\item{lambda.lim}{If not \code{NULL} gives a fixed value for \eqn{\lambda_{lim}}. Overrides any value passed by \code{phi}.}

\item{rm.COCs}{If \code{TRUE} (default), automatically removes \emph{"compound oulier clusters" (COCs)}.}

\item{extract}{If \code{TRUE} (default), returns a list of cluster names. If \code{FALSE} only prints a tree with no \code{outlier clusters} indicated.}

\item{plot.tree}{If \code{TRUE} (default), plots tree. Has no effect if \code{extract=FALSE}.}

\item{plot.graph}{If \code{TRUE} (default), plots all \eqn{\lambda} ordered from low to high and indicates which values are above \eqn{\lambda_{lim}}.}

\item{LD_threshold1}{Extracts clusters that have a minimum pair wise LD value above this threshold. Used for LD binning (clustering method for LDnaRAW shold be other than 'single').}

\item{LD_threshold2}{Extracts clusters that have a mean pair wise LD value above this threshold. Used for LD binning (clustering method for LDnaRAW shold be other than 'single').}

\item{extract.fun}{Only used internally}
}
\value{
If extract=TRUE a named list of vectors giving the loci for the extracted clusters; otherwise returns plots of the tree (if plot.tree=TRUE) and the distribution of \eqn{\lambda} values (if plot.graph=TRUE and extract=TRUE)
}
\description{
Identifies \emph{outlier clusters}, \emph{OCs}, and plots \code{"single linage clustering trees"} that describes cluster merging with decreasing LD threshold.
}
\details{
If \code{plot.tree} and \code{plot.graph} are set to \code{TRUE}, \code{extractClusters} plots two graphs (default). The first shows all \eqn{\lambda}-values oredered from low to high and indicates which values are above \eqn{\lambda_{lim}}. Values corresponding to \emph{"selected outlier clusters", SOCs} are indicated in red and values corresponding to \emph{COCs} are indiced in blue. A \emph{COC} is defined as any \emph{OC} that contains loci from an \emph{OC} already extracted at a higher LD threshold. The second graph gives the tree illustrating cluster merger with decreasing LD threshold where nodes represent and node distance gives the LD threholds at which these events occur. Branches corresponding to \emph{SOCs} are indicated in red and those corresponding to \emph{COCs} are indiced in blue (if \code{rm.COCs=FALSE}).
}
\examples{
# Simple upper diagonal LD matrix
LDmat <- structure(c(NA, 0.84, 0.64, 0.24, 0.2, 0.16, 0.44, 0.44, NA, NA, 0.8, 0.28, 0.4, 0.36, 0.36, 0.24, NA, NA, NA, 0.48, 0.32, 0.2, 0.36, 0.2, NA, NA, NA, NA, 0.76, 0.56, 0.6, 0.2, NA, NA, NA, NA, NA, 0.72, 0.68, 0.24, NA, NA, NA, NA, NA, NA, 0.44, 0.24, NA, NA, NA, NA, NA, NA, NA, 0.2, NA, NA, NA, NA, NA, NA, NA, NA), .Dim = c(8L, 8L), .Dimnames = list(c("L1", "L2", "L3", "L4", "L5", "L6", "L7", "L8"), c("L1", "L2", "L3", "L4", "L5", "L6", "L7", "L8")))
# Calculate raw data
ldna <- LDnaRaw(LDmat)
# Extract clusters and plot graphs, for this example min.edges=0 such that each tip clade corresponds to an individual locus
par(mfcol=c(1,2))
clusters <- extractClusters(ldna, min.edges=0, phi=1)
clusters <- extractClusters(ldna, min.edges=0, phi=0.25, rm.COCs=FALSE)
# Larger data set
data(LDna)
ldna <- LDnaRaw(r2.baimaii_subs)
# Only print trees
clusters <- extractClusters(ldna, min.edges=15, extract=FALSE)
clusters <- extractClusters(ldna, min.edges=5, extract=FALSE)
# Different values of min.edges and phi can have crucial effects which clusters are extracted
clusters <- extractClusters(ldna, min.edges=15, phi=3)
clusters <- extractClusters(ldna, min.edges=10, phi=2)
# Include COCs
clusters <- extractClusters(ldna, min.edges=15, phi=5, rm.COCs=FALSE)
# Extract clusers without plottin graphs
clusters <- extractClusters(ldna, min.edges=15, phi=5, plot.tree=FALSE, plot.graph=FALSE)
# Set fixed value for lambda.lim
clusters <- extractClusters(ldna, min.edges=15, lambda.lim=2)
}
\seealso{
\code{\link{LDnaRaw}}, \code{\link{summaryLDna}} and \code{\link{plotLDnetwork}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, Christopher Knight \email{Chris.Knight@manchester.ac.uk}
}
\keyword{extractClusters}
