% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDnClusteringEL.R
\name{LDnClusteringEL}
\alias{LDnClusteringEL}
\title{Linkage disequilibrium (LD) network clustering, starting from edge lists of pairwise LD-values as input.}
\usage{
LDnClusteringEL(
  EL_path = "./LD_EL",
  nSNPs = 1000,
  w1 = 10,
  columns = c(1, 2, 4),
  out_folder = "LDnCl_out",
  min_LD = 0.7,
  plot.network = NULL,
  threshold_net = 0.9,
  to_do = NULL,
  cores = 1,
  min.cl.size = 2,
  min_LD2 = 0.3
)
}
\arguments{
\item{EL_path}{Path to folder (only) containing relevant el's (one per chromosome/linkage group). For convenience the file name should be "name_of_chromosome"."el" e.g. chr1.el, 1.el, LG1.el, LG1_clean.el etc.}

\item{nSNPs}{Desired number of SNPs per window.}

\item{columns}{Index of the column that contain LD (r^2). The two first two columns must be locus 1 and 2 for each edge.}

\item{out_folder}{Path to folder where output is produced (default is './LDnCl_out/'). Proceed to use \code{Concat_files} to concatenate this data to a single file.}

\item{min_LD}{Minimum LD value that at least one locus must be connected to all other loci in the recursive step.}

\item{plot.network}{File name for plotting network. If \code{NULL} (default) no network is plotted.}

\item{threshold_net}{Threshold for edges when plotting network.}

\item{to_do}{Vector with indexes for files in folder specified by \code{EL_path} that still need to be done. If \code{NULL} (default) all files will be processed.}

\item{cores}{Number of cores, default is 1}

\item{min.cl.size}{If 1 also singletons will be retained, which may be necessary for data sets with weak LD-structures (e.g. most loci are independent). Produced very large files though, so typically 2 is used to exclude all singleton clusters.}

\item{min_LD2}{Window breakpoints are placed where r^2 between adjacent SNPs within a window (defined by w1) is below min_LD2.}
}
\value{
Returns a list of three objects which are saved as ".rds" files in the folder specified by \code{out_folder}.
\code{cluster_summary} is a data frame that contains most of the relevant information for each cluster, including summary statistics (Median LD, MAD etc, see below). 
\code{clusters} contains a list of locus names (chr:position; each entry corresponding to a row in \code{cluster_summary}.
\code{MCL} is a vector of names which which best represents the LD-cluster in downstream analyses ('maximally connected SNP', MCL aka rSNP).
\cr 
\cr
Each .rds file contains only information for each chromosome but they can be concatenated into a single file using \code{\link{Concat_files}}. 
\cr
\cr
The columns in file \code{cluster_summary} are:
Chr', 'Window', 'Pos', 'Min', 'Max', 'Range', 'nSNPs', 'Min_LD'
\item{Chr}{Chromosome or linkage group identifer}
\item{Window}{Window identifier, recycled among chromosomes}
\item{Pos}{Mean position of SNPs in a cluster}
\item{Min}{Most downstream position of SNPs in a cluster}
\item{Max}{Most upstream position of SNPs in a cluster}
\item{Range}{Max-Min}
\item{nSNPs}{Number of SNPs in the cluster}
\item{Min_LD}{the minimum LD between the rSNP/MCL and all other loci in its cluster}
}
\description{
Finds clusters of loci connected by high LD within non-overlapping windows and returns a summary file, the resulting clusters and the names of the rSNPs (previouslu MCL), to be use e.g. in Genome Wide Association (GWA) analyses.
}
\details{
Uses single linkage clustering within non-overlapping windows of SNPs (within chromosomes) to find groups of correlated SNPs connected by high LD. This is done recursively within the single linkage clustering sub-trees (from root and up); as soon as a clade is reached where at least one locus is connected with all other loci within its clade above threshold \code{min_LD}, the algorithm stops. The default (0.7) produces clusters where the first PC typically explains >99 percent of the variation in each cluster.

Uses edge lists of LD values as input. The informative columns are specified by \code{LD_column}, where first and second specifies the columns for locus names and the third contains LD (r^2) values.
\cr
\cr
Window breakpoints are placed where r^2 between adjacent SNPs within a window (defined by \code{w1}) is below \code{min_LD2}.
\cr
\cr
\code{nSNPs} determines the goal size for each window; smaller windows will produce faster computation times. \cr
\cr
\cr
If something goes wrong you can specify \code{to_do} with indexes for files in folder specified by \code{EL_path} that still need to be done. 
\cr
\cr
It is also possible to plot LD-networks for different cluster solutions. For this you can specify a path for where the figure will be saved (\code{plot.network}). 
The parameter \code{threshold_net} determines the minimum value for each edge that is allowed in the network (i.e is equivalent to a single linkage clustering network). 
Each LD-cluster has a unique (arbitrary) color combination and black vertices represent loci that are not part of any cluster, i.e. should be considered independently in GWA analyses. 
Minimum for \code{plot.network} is './', which exports figure to the working directory. Only works for data sets with up to 1000 SNPs and it is recommended that \code{threshold_net} is ~LD2.
}
\examples{

}
\references{
Kemppainen, P., Knight, C. G., Sarma, D. K., Hlaing, T., Prakash, A., Maung Maung, Y. N., Walton, C. (2015). Linkage disequilibrium network analysis (LDna) gives a global view of chromosomal inversions, local adaptation and geographic structure. Molecular Ecology Resources, 15(5), 1031-1045. https://doi.org/10.1111/1755-0998.12369\cr
\cr
Li, Z., Kemppainen, P., Rastas, P., Merila, J. Linkage disequilibrium clustering-based approach for association mapping with tightly linked genome-wide data. Accepted to Molecular Ecology Resources.
}
\seealso{
\code{\link{emmax_group}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}, zitong.li \email{lizitong1985@gmail.com}
}
\keyword{LD,}
\keyword{Linkage}
\keyword{analyses,complexity}
\keyword{disequilibrium,}
\keyword{network}
\keyword{reduction}
