% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDnetwork.r
\name{plotLDnetwork}
\alias{plotLDnetwork}
\title{Plots linkage disequilibrium networks}
\usage{
plotLDnetwork(
  ldna,
  LDmat,
  option,
  threshold,
  clusters,
  summary,
  digits = 2,
  exl = NULL,
  full.network = TRUE,
  include.parent = FALSE,
  after.merger = FALSE,
  graph.object = FALSE,
  col = "grey",
  frame.color = "grey",
  pos = NULL
)
}
\arguments{
\item{ldna}{Output from \code{\link{LDnaRaw}}}

\item{LDmat}{a matrix of pairwise LD values}

\item{option}{\code{option=1} prints a full LD network from and edge list (\code{LDmat}) and a specified LD threshold (\code{threshold}). \code{option=2} prints LD networks based on the files \code{clusters} and \code{summary} that contains information of extracted clusters.}

\item{threshold}{Specifies the LD threshold at which an LD network is plotted. Only required for option=1.}

\item{clusters}{Output from \code{\link{extractClusters}}.}

\item{summary}{Output from \code{\link{summaryLDna}}.}

\item{digits}{Needs to be the same as used \code{\link{LDnaRaw}} and \code{\link{extractClusters}}, if not default (2)}

\item{exl}{A list of locus names to be excluded from the LD networks (default is \code{NULL}).}

\item{full.network}{If \code{TRUE} (default), includes all loci in the LD networks, not recommended for large data sets.}

\item{include.parent}{If \code{full.network=FALSE} and \code{include.parent=TRUE} all loci from the parent cluster (after merger) are included. If \code{include.parent=FALSE} only the focal cluster is shown.}

\item{after.merger}{Whether to show LD networks at an LD threshold just before (\code{FALSE}) or just after (\code{TRUE}) merger.}

\item{graph.object}{Whether to output \code{graph object} when \code{option=1} (default=\code{FALSE}).}

\item{col}{Color of vertices when using \code{option=1}, default is "grey"}

\item{frame.color}{Used by \code{\link{LDnClustering}}}

\item{pos}{A numeric vector giving the position of loci along each chromosome. This is converted into red-green color space such that within each cluster it is possible to infer if vertex position reflexts its physical position in the chromosome. Currently works only for \code{option=1}.}
}
\value{
If \code{option=1} and \code{graph.object=TRUE} the output is an igraph.object that can further be manipulated for custom networks (see \code{\link{igraph}} for details).
}
\description{
Allows for visual inspection of clusters identified by \code{\link{extractClusters}}.
}
\details{
See examples for more details
}
\examples{
### Example with option=1
data(LDna)
plotLDnetwork(LDmat=r2.baimaii_subs, option=1, threshold=0.4)
plotLDnetwork(LDmat=r2.baimaii_subs, option=1, threshold=0.4, col="red") # for more color
### Examples with option 2
par(mfcol=c(1,2))
ldna <- LDnaRaw(r2.baimaii_subs)
clusters <- extractClusters(ldna, r2.baimaii_subs, min.edges=20, lambda.lim = 1)
summary <- summaryLDna(ldna, clusters,r2.baimaii_subs)
#default settings with option=2
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]], summary=summary)
## Other useful settings
# For large data sets
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]], summary=summary, full.network=FALSE, include.parent=FALSE, after.merger=FALSE)
# To visualise the merger
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]], summary=summary, full.network=TRUE, after.merger=TRUE)
# Or
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]], summary=summary, full.network=FALSE, include.parent=TRUE, after.merger=TRUE)
# To show that ususally several clusters are involved in most mergers
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]], summary=summary, full.network=FALSE, include.parent=TRUE, after.merger=FALSE)
### Print directly to file, recommended for large data sets with many clusters. 
\dontrun{
library(parallel)
fun <- function(x){
setEPS()
postscript(paste(x, "network.eps",  sep="_"))
plotLDnetwork(ldna, r2.baimaii_subs, option=2, clusters=clusters[[1]][x], summary=summary[x,], full.network=FALSE)
dev.off()
}
lapply(1:length(clusters), fun)
# a multicore version of this
mclapply(1:length(clusters), fun, mc.cores=4, mc.preschedule=TRUE)}
}
\seealso{
\code{\link{LDnaRaw}}, \code{\link{extractClusters}} and \code{\link{summaryLDna}}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}
\keyword{plotLDnetwork}
