% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r2.baimaii_subs.R
\docType{data}
\name{r2.baimaii_subs}
\alias{r2.baimaii_subs}
\title{Matrix of pairwise \eqn{R^2} values}
\usage{
r2.baimaii_subs
}
\description{
Lower diagonal pairwise matrix of \eqn{R^2} values between 957 SNPs from 91 wild-caught individuals of the SE Asian mosquito malaria vector, \emph{Anopheles baimaii}.
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}

